Move Enclosure, File and Gzdecode classes to namespace
